+++ lib/firestore_service.dart
@@ -2,11 +2,13 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 
 class FirestoreServicePackage<T> {
   final String collectionName;
+  final String searchBy;
   final List Function(QuerySnapshot) dataListFromSnapshot;
   final int limitOfRetrievedData;
 
   FirestoreServicePackage(
       {this.collectionName,
+      this.searchBy,
       this.dataListFromSnapshot,
       this.limitOfRetrievedData});
   final FirebaseFirestore firebasefirestore = FirebaseFirestore.instance;
@@ -15,7 +17,7 @@ class FirestoreServicePackage<T> {
     final collectionReference = firebasefirestore.collection(collectionName);
 
     return collectionReference
-        .where('username', isGreaterThanOrEqualTo: query ?? '')
+        .where('${searchBy ?? ''}', isGreaterThanOrEqualTo: query ?? '')
         .limit(limitOfRetrievedData)
         .snapshots()
         .map(dataListFromSnapshot);